[#17232] CMake: Use PRIVATE linking between test-app and LLVM targets
Without the PRIVATE, libraries linked to the LLVM target are propagated
by the CMake to every target linking LLVM. This is the behaviour of the
PUBLIC, which seems to be the default when not specifying any of those
while linking.

We don't want that, because then we can't be sure LLVM library is all we
need to disassembly (test-app could take some missing functions from the
LLVM libraries).

Earlier we used some hacks to achieve that, but the use of PRIVATE is
much cleaner. The only caveat is, if any target_link_libraries command
specify PRIVATE/PUBLIC/INTERFACE, all such commands need to specify one
of those for the given target.