[#17232] test-app: Don't disassemble zeros added by the strtoull
"Bytes" argument is internally converted to uint64_t, so even if only
0x45 is the argument, it's being completed with zeros to make it 8B.

Previously, the size of the variable (i.e. 8B) was passed to the LLVM,
which made it translate also those additional zeros. For some
architectures, zeros translate to valid instructions (e.g. 0x0000 is
movs r0, r0 on ARM's thumb).

After the change, LLVM will only translate bytes passed as the argument
(with at most 1 additional zero if the number of HEX digits is uneven).