[#31629] CMake: Build LLVM libs from source by default if missing
The code for getting libraries from binary packages was removed.

It was tough to maintain in case of LLVM upgrades. Packages for all the
used architectures aren't available for every release.

However, it's still possible to use prebuilt libraries. If all of the
required libraries (listed in `LLVM_LIBS`) are found in the
`${LLVM_LIBS_DIR}`, they are used to build `libllvm-disas`.

The libraries are copied to `${LLVM_LIBS_DIR}` after they're built so
removing a build directory won't require LLVM libraries to be rebuilt.